apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
redis==5.1.1
apscheduler==3.10.4
openai==1.51.0
PyYAML==6.0.2
sqlglot==25.24.4
python-dotenv==1.0.1
structlog==24.4.0
httpx==0.27.2
python-multipart
